Make the config of WindowState synced with the client side
The original logic may cause WMS to skip sending the configuration to
the client side unexpectedly. For example, a portrait app is resuming
from a landscape app, this portrait app may get the landscape config
from relayoutWindow() first. However, WMS would not change the config
of WindowState to the one returned to the client side. When AMS sets
the new config (the portrait one) to WMS, WMS won't report the new
config to the portrait app, because the config of the WindowState and
the config of WMS may be the same (portrait). The latest config from
onConfigurationChanged() of the portrait app would be landscape... :(

Note: performLayoutAndPlaceSurfacesLockedInner() would be skipped if
mWaitingForConfig is true, that makes WMS skip setting the server
config to the WindowState in relayoutWindow().

Change-Id: Ifb08b1e7b0ed5fda8fba5c7253f496c55665580e
Signed-off-by: mydongistiny <jaysonedson@gmail.com>