Catch the runtime exception thrown by a bad backup transport
Symptom:
BackupManagerService was crashed by NPE thrown
from a bad application's backup transport.

Root cause:
RuntimeException is not caught on the restore operation.
When a backup transport throw a runtime exception, the service is
crashed with unhandled exception.
There is no fail safe to avoid the system crash.

Solution:
Catch the runtime exception, thrown by a bad backup transport.
The package is not restored. It's just skipped.

Change-Id: Idd768b41df586cb768197e69736c5528851dc98e