OMS: Fix Quick settings tiles font themes derp [6/6]
KreAch3R:
Picked for our substratum related font implementation

This fixes quick settings tiles not changing when you apply a font via
a theme. I didn't notice this until Randall mentioned it and I tried it
with GRVY and sure enough.

Anyways, this mirrors the way we do things with the keyguard and every
other element in the ROM where the font changes.

Change-Id: I03279f2e1cc95fdcb68400ea28d6193df96bff1e